Traffic Signs Classification/requirements.txt
numpy==1.24.3
pandas==2.0.3
matplotlib==3.7.2
tensorflow==2.12.0
scikit-learn==1.3.0
pillow==10.0.0
opencv-python==4.8.0.76 